subprocess.py: ensure SIGPIPE is honored by sub-processes
This is a workaround for a nasty issue when using
the standard python subprocess module, or os.exec* etc.
in any version of python before 3.2
http://bugs.python.org/issue1652

It's disappointing that a simple patch was provided
to python for this important issue _5 years ago_ and
it's only made it into python 3.2, with not much chance
of upstream updating earlier releases.